Update README.md
Because the wiki migration will be happening soon, the short links that are not directly on the wiki will not get updated.  We were unaware that we were ever going to run into this issue, and following the migration, it will still be the correct way to get the URL.  However, the redirect will work for the full URL because the name of the page has not changed.  We should update all of the URLs on GitHub again once the wiki migration is complete.

Co-authored-by: Herbert Damker <herbert.damker@telekom.de>